+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.0.4 (2017-10-04)
+- fix security issue with 2FA where any YubiKey OTP would be accepted to 
+  connect to VPN service when user was not enrolled for YubiKey 2FA
+
 ## 1.0.3 (2017-09-19)
 - increase TOTP attempt count to 60 from 10 per hour (issue #64)
 - fix source formatting and method annotations